Edit ReadMe
A joke is a display of humour in which words are used within a specific and well-defined narrative structure to make people laugh. It takes the form of a story, usually with dialogue, and ends in a punch line. It is in the punch line that the audience becomes aware that the story contains a second, conflicting meaning. This can be done using a pun or other word play such as irony, a logical incompatibility, nonsense or other means.